+++ pkg/test/RabbitmqStompExtension.php
@@ -8,9 +8,14 @@ use PHPUnit\Framework\SkippedTestError;
 
 trait RabbitmqStompExtension
 {
+    private function getDsn()
+    {
+        return getenv('RABITMQ_STOMP_DSN');
+    }
+
     private function buildStompContext(): StompContext
     {
-        if (false == $dsn = getenv('RABITMQ_STOMP_DSN')) {
+        if (false == $dsn = $this->getDsn()) {
             throw new SkippedTestError('Functional tests are not allowed in this environment');
         }
